1707: Refactored common basic auth handling code from report-server and entity-server into @tupai/auth
